refactor: app모듈 패키지 구조 수정
